requirements.txt
absl-py==2.0.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.4
contourpy==1.1.1
cycler==0.12.0
cymem==2.0.8
Cython==3.0.6
dash==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
et-xmlfile==1.1.0
executing==2.0.0
fastexcel==0.14.0
fastjsonschema==2.19.0
filelock==3.13.1
flake8==6.1.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.43.0
fqdn==1.5.1
funpymodeling==0.1.8
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googletrans==4.0.0rc1
gower==0.1.2
grpcio==1.59.2
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.5.31
idna==2.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
keras==2.15.0
kiwisolver==1.4.5
kmodes==0.12.2
langcodes==3.3.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
narwhals==1.34.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numdifftools==0.9.41
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==6.0.1
pluggy==1.3.0
pmdarima==2.0.4
polars==1.29.0
pre-commit==3.5.0
preshed==3.0.9
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.9.1
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.13.0
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
semopy==2.3.11
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post9
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.13.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.24.6
wasabi==1.1.2
watchdog==6.0.0
wcwidth==0.2.8
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wordcloud==1.9.2
wrapt==1.14.1
XlsxWriter==3.2.0
zipp==3.21.0
